Refactor UI components and add analytics data fetch functionality
This commit refactored some UI components to improve readability and maintainability of the code. It also included a new functionality to fetch analytics data for Woocommerce stores. The changes included refactoring dashboard and stores attributes, and adding two new components for modularization of the dashboard, i.e., all-stores-dashboard and store-dashboard.

Signed-off-by: Rahul Retnan <rahul@raftlabs.com>