Implement parameter encryption in route management
The changes introduce encryption and decryption of parameters in store route management. The route params for 'storeId' are now encrypted using 'btoa()' on generation and decoded with 'atob()' on extraction. Also, breadcrumb and sidebar navigation were updated to reflect encrypted paths and 'url-validator.ts' was removed as it's no longer used.

Signed-off-by: Rahul Retnan <rahul@raftlabs.com>